refactor(tee): split tee/verify.rs god object into per-backend modules
tee/verify.rs (884 LOC, the tee-verify quote-verification impl) -> tee/verify/
by attestation backend: tdx (Intel TDX), sev_snp (AMD SEV-SNP; module named
sev_snp to avoid shadowing the extern `sev` crate), certs (X.509 validity +
CRL revocation), nitro (AWS Nitro). mod.rs keeps the shared VerifiedQuote type,
the verify_quote_signature dispatcher, evidence parsing, consts, and the
cross-module tests_now_from_collateral helper. Tests move to verify/tests.rs
(include_bytes paths repointed one level up). Pure code motion, no behavior
change; max production module 176 LOC.
