+++ scripts/.god-object-baseline.txt
@@ -8,7 +8,6 @@
 1065 sandbox-runtime/src/session_auth.rs
 948 ai-agent-instance-blueprint-lib/src/workflows.rs
 922 ui/src/routes/workflows._index.tsx
-884 sandbox-runtime/src/tee/verify.rs
 838 sandbox-runtime/src/reaper.rs
 744 sandbox-runtime/src/scoped_session_auth.rs
 721 ui/src/components/shared/SessionSidebar.tsx
